requirements.txt
Flask==3.1.3
pymysql==1.1.2
python-dotenv==1.0.1
Werkzeug==3.1.6
gunicorn==21.2.0
